refactor: extract pytest fixtures out of api/conftest.py into api/tests/fixtures.py
The api conftest.py is shipped as a top-level `py-module` in the installed
wheel. Downstream consumers that install the api into their venv could not
re-use its fixtures via `from conftest import *` because their own root
`conftest.py` collides with the api's top-level `conftest` module.

Move every `@pytest.fixture` (and the helper constants they reference)
out of `api/conftest.py` into a regular module at `api/tests/fixtures.py`,
and have `api/conftest.py` re-export them with `from tests.fixtures import *`.

Pytest hooks (`pytest_addoption`, `pytest_configure`) stay in `api/conftest.py`
because pytest only invokes hooks from conftest / registered plugin modules.

Downstream consumers can now write:

    # consumer's conftest.py
    from tests.fixtures import *  # noqa: F401, F403

without any importlib shenanigans.
